♻️ move indexada_exceptions to www_som
