+++ www_som/exceptions/__init__.py
