+++ www_som/exceptions/indexada_exceptions.py
